Refactored tile to use u8 internally instead of f32 and optimised system so that it only syncs changed positions

+++ src/game/tile_map/factories/map.rs
@@ -11,7 +11,7 @@ pub const MAP_HEIGHT: f32 = 32.;
 pub fn initialise_map(world: &mut World, sprite_store: &SpriteStore) {
     for y in 0..(MAP_HEIGHT as usize) {
         for x in 0..(MAP_WIDTH as usize) {
-            initialise_tile(world, x as f32, y as f32, sprite_store.get_ground());
+            initialise_tile(world, x as u8, y as u8, sprite_store.get_ground());
         }
     }
 }
